step_0_data/data_requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
biopython=1.83=py312h98912ed_0
boost-cpp=1.70.0=h8e57a91_2
brotli-python=1.1.0=py312h30efb56_1
bzip2=1.0.8=hd590300_5
ca-certificates=2024.2.2=hbcca054_0
certifi=2024.2.2=pyhd8ed1ab_0
charset-normalizer=3.3.2=pyhd8ed1ab_0
conterminator=1.c74b5=h95f258a_1
diamond=2.0.4=h56fc30b_0
gawk=5.3.0=ha916aea_0
gettext=0.21.1=h27087fc_0
gmp=6.3.0=h59595ed_0
icu=64.2=he1b5a44_1
idna=3.6=pyhd8ed1ab_0
ld_impl_linux-64=2.40=h41732ed_0
libblas=3.9.0=21_linux64_openblas
libcblas=3.9.0=21_linux64_openblas
libexpat=2.5.0=hcb278e6_1
libffi=3.4.2=h7f98852_5
libgcc-ng=13.2.0=h807b86a_5
libgfortran-ng=13.2.0=h69a702a_5
libgfortran5=13.2.0=ha4646dd_5
libgomp=13.2.0=h807b86a_5
liblapack=3.9.0=21_linux64_openblas
libnsl=2.0.1=hd590300_0
libopenblas=0.3.26=pthreads_h413a1c8_0
libsqlite=3.44.2=h2797004_0
libstdcxx-ng=13.2.0=h7e041cc_5
libuuid=2.38.1=h0b41bf4_0
libxcrypt=4.4.36=hd590300_1
libzlib=1.2.13=hd590300_5
llvm-openmp=8.0.1=hc9558a2_0
mpfr=4.2.1=h9458935_0
ncurses=6.4=h59595ed_2
numpy=1.26.4=py312heda63a1_0
openmp=8.0.1=0
openssl=3.2.1=hd590300_0
packaging=23.2=pyhd8ed1ab_0
pandas=2.2.0=py312hfb8ada1_0
pip=24.0=pyhd8ed1ab_0
plotly=5.18.0=pyhd8ed1ab_0
prodigal=2.6.3=h031d066_7
pysocks=1.7.1=pyha2e5f31_6
python=3.12.1=hab00c5b_1_cpython
python-dateutil=2.8.2=pyhd8ed1ab_0
python-tzdata=2023.4=pyhd8ed1ab_0
python_abi=3.12=4_cp312
pytz=2024.1=pyhd8ed1ab_0
readline=8.2=h8228510_1
requests=2.31.0=pyhd8ed1ab_0
scipy=1.12.0=py312heda63a1_2
setuptools=69.0.3=pyhd8ed1ab_0
six=1.16.0=pyh6c4a22f_0
tenacity=8.2.3=pyhd8ed1ab_0
tk=8.6.13=noxft_h4845f30_101
tzdata=2024a=h0c530f3_0
urllib3=2.2.0=pyhd8ed1ab_0
wheel=0.42.0=pyhd8ed1ab_0
xz=5.2.6=h166bdaf_0
zlib=1.2.13=hd590300_5
